# DentalERP/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dental ERP Authentication API
  version: 1.0.0
  description: |
    Authentication contract using Laravel Sanctum and Spatie Permission.

    Lifecycle governance follows Accepted DD-AUTH-007 and Accepted ADR-005.
    Field classification, exposure, nullability, and field governance follow
    Accepted DD-AUTH-017. DD-AUTH-005 remains superseded historical evidence.

    All Session and Device operations are self-service. Super Admin uses the same
    ownership boundary. Cross-user administration is outside this contract under
    Accepted DD-AUTH-003.

    Token TTL:
    - Access Token: 60 minutes.
    - Refresh Token: 30 days.
    - Password Reset Token: 15 minutes.
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /auth/logout-all:
    post:
      summary: Log out all sessions
      description: 'AUTH-BR-009: Revokes all active User Sessions owned by the authenticated User and every descendant token. Devices remain registered, Login History Operational History Projections are retained, and canonical Audit Events remain immutable.'
      operationId: auth.logoutAll
      x-business-rules: [AUTH-BR-009]
      x-decisions: [ADR-002, ADR-006, DD-AUTH-001, DD-AUTH-003, DD-AUTH-007, DD-AUTH-017, ADR-005]
      tags: [Authentication]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: All sessions were revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              examples:
                success:
                  value: { success: true, message: 'All sessions revoked successfully.', data: null, errors: null, meta: null }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/refresh:
    post:
      summary: Rotate a refresh token
      description: |
        Rotates Refresh Token within the current User Session and revalidates
        Session, User, Organization, Branch, and Device state.

        Access Token expires in 60 minutes.

        Refresh Token expires in 30 days.

        Old Refresh Token becomes invalid immediately.

        Reuse of a rotated Refresh Token revokes its entire token family, owning
        Session, and descendant Access Token under DD-AUTH-007.

        Rotation remains within the same Refresh Token family and User Session. The
        previous active Access Token and Refresh Token become invalid before the
        replacement pair becomes active.
      operationId: auth.refresh
      x-business-rules: [AUTH-BR-002, AUTH-BR-013]
      x-decisions: [ADR-002, DD-AUTH-001, DD-AUTH-007, DD-AUTH-017, ADR-005]
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
            examples:
              rotate: { value: { refresh_token: refresh-token-placeholder-0000000000000001 } }
      responses:
        '200':
          description: Token pair rotated successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPairResponse' }
              examples:
                success: { $ref: '#/components/examples/RefreshSuccess' }
        '401': { $ref: '#/components/responses/InvalidRefreshToken' }
        '403': { $ref: '#/components/responses/InactiveTenant' }
        '409': { $ref: '#/components/responses/TokenReuseDetected' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/forgot-password:
    post:
      summary: Request password reset instructions
      description: |
        Queues reset instructions when an account exists and always returns the same
        accepted response to prevent enumeration.

        Password Reset Token expires in 15 minutes and can be used only once.
      operationId: auth.forgotPassword
      x-business-rules: [AUTH-BR-010]
      x-decisions: [ADR-003, ADR-006, DD-AUTH-007, DD-AUTH-017, ADR-005]
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
            examples:
              requestReset: { value: { email: doctor@example.com } }
      responses:
        '202':
          description: Reset request accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              examples:
                accepted:
                  value: { success: true, message: 'If the email is registered, password reset instructions will be sent.', data: null, errors: null, meta: null }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/reset-password:
    post:
      summary: Reset a forgotten password
      description: |
        Resets the password using a single-use Password Reset Token, revokes every
        active User Session and all descendant Access Tokens and Refresh Token families,
        preserves registered Devices, and records an audit event.

        Password Reset Token is valid for 15 minutes and can be used only once.
      operationId: auth.resetPassword
      x-business-rules: [AUTH-BR-011]
      x-decisions: [ADR-003, ADR-006, DD-AUTH-007, DD-AUTH-017, ADR-005]
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
            examples:
              reset:
                value:
                  email: doctor@example.com
                  token: reset-token-placeholder
                  password: NewStrongPassword456!
                  password_confirmation: NewStrongPassword456!
      responses:
        '200':
          description: Password reset and sessions revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              examples:
                success:
                  value: { success: true, message: 'Password reset successfully. Please log in again.', data: null, errors: null, meta: null }
        '400': { $ref: '#/components/responses/InvalidResetToken' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/change-password:
    post:
      summary: Change the authenticated user's password
      description: |
        Verifies the current password and changes it according to the approved policy.

        Current Session and its active Access/Refresh Token pair remain active.

        Every other active Session is revoked together with descendant Access Tokens
        and Refresh Token families.

        Registered Devices remain registered. An immutable PASSWORD_CHANGED audit event
        is generated, and Login History remains unchanged. DD-AUTH-018 defines
        credential-change revocation as an exception to the generic logout_at trigger.
      operationId: auth.changePassword
      x-business-rules: [AUTH-BR-012, AUTH-BR-016]
      x-decisions: [ADR-006, DD-AUTH-004, DD-AUTH-007, DD-AUTH-017, DD-AUTH-018, ADR-005]
      tags: [Authentication]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
            examples:
              change:
                value:
                  current_password: StrongPassword123!
                  password: NewStrongPassword456!
                  password_confirmation: NewStrongPassword456!
      responses:
        '200':
          description: Password changed successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangePasswordResponse' }
              examples:
                success:
                  value:
                    success: true
                    message: Password changed successfully.
                    data:
                      current_session_active: true
                      other_sessions_revoked: true
                      registered_devices_retained: true
                    errors: null
                    meta: null
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 32
          writeOnly: true
          description: Single-use opaque Refresh Token with a 30-day TTL.
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 150 }
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [email, token, password, password_confirmation]
      properties:
        email: { type: string, format: email, maxLength: 150 }
        token:
          type: string
          writeOnly: true
          description: Single-use Password Reset Token with a 15-minute TTL.
        password: { type: string, format: password, minLength: 8, writeOnly: true }
        password_confirmation: { type: string, format: password, minLength: 8, writeOnly: true }
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [current_password, password, password_confirmation]
      properties:
        current_password: { type: string, format: password, writeOnly: true }
        password: { type: string, format: password, minLength: 8, writeOnly: true }
        password_confirmation: { type: string, format: password, minLength: 8, writeOnly: true }
...
    TokenPair:
      type: object
      required: [token_type, access_token, access_token_expires_at, refresh_token, refresh_token_expires_at]
      properties:
        token_type: { type: string, const: Bearer }
        access_token: { type: string, description: Opaque Access Token returned only at issuance or rotation. }
        access_token_expires_at:
          type: string
          format: date-time
          description: Access Token expiry timestamp; default lifetime is 60 minutes.
        refresh_token: { type: string, description: Opaque Refresh Token returned only at issuance or rotation. }
        refresh_token_expires_at:
          type: string
          format: date-time
          description: Refresh Token expiry timestamp; default lifetime is 30 days.
...
    MessageResponse:
      type: object
      required: [success, message, data, errors, meta]
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { type: ['null'] }
        errors: { type: ['null', object] }
        meta: { type: ['null', object] }
    ChangePasswordResponse:
      type: object
      required: [success, message, data, errors, meta]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [current_session_active, other_sessions_revoked, registered_devices_retained]
          properties:
            current_session_active: { type: boolean, const: true }
            other_sessions_revoked: { type: boolean, const: true }
            registered_devices_retained: { type: boolean, const: true }
        errors: { type: ['null'] }
        meta: { type: ['null'] }
    TokenPairResponse:
      type: object
      required: [success, message, data, errors, meta]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/TokenPair' }
        errors: { type: ['null'] }
        meta: { type: ['null'] }
...
    ErrorResponse:
      type: object
      required: [success, message, data, errors, meta]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        data: { type: ['null'] }
        errors: { type: ['null', object] }
        meta: { type: ['null', object] }
...
  responses:
...
    Unauthenticated:
      description: Missing, invalid, expired, or revoked access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Unauthenticated.', data: null, errors: null, meta: null }
...
    InactiveTenant:
      description: User, Organization, Branch, or device is inactive, revoked, or mismatched.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Authentication context is not active.', data: null, errors: null, meta: null }
    InvalidRefreshToken:
      description: Refresh token is invalid, expired, or revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Invalid refresh token.', data: null, errors: null, meta: null }
    InvalidResetToken:
      description: Password reset token is invalid, expired, or already used.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Invalid or expired reset token.', data: null, errors: null, meta: null }
    TokenReuseDetected:
      description: A rotated refresh token was reused; its entire token family, owning Session, and descendant Access Token were revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Refresh token reuse detected.', data: null, errors: null, meta: null }
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            message: The given data was invalid.
            data: null
            errors:
              email: ['Email must be valid.']
            meta: null
...
    TooManyRequests:
      description: Authentication operation was rate-limited.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Too many requests.', data: null, errors: null, meta: { retry_after: 60 } }
    ServerError:
      description: Unexpected server failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { success: false, message: 'Internal server error.', data: null, errors: null, meta: null }
...
  examples:
...
    RefreshSuccess:
      value:
        success: true
        message: Token refreshed successfully.
        data:
          token_type: Bearer
          access_token: 2|new-access-token-placeholder-000000000000001
          access_token_expires_at: '2026-08-02T15:46:00Z'
          refresh_token: new-refresh-token-placeholder-000000000000001
          refresh_token_expires_at: '2026-09-01T14:46:00Z'
        errors: null
        meta: null
...